I messed up with namespace
Basically, i messed up when trying to recreate this project. I forgot to commit it originally and had continued on despite that fact. During my recreation of this version I changed the base name of the project and couldn't figure out a way to alter it. This should be the fix for that problem however it could also not work... I know how to code but sometimes things just don't work out how you think they will.